🔧 [#43] chore: 패키지 이동 domain/recruite -> domain/recruit/entity
