+++ src/main/java/gaji/service/domain/recruit/entity/SearchKeyword.java
@@ -1,0 +1,28 @@
+package gaji.service.domain.recruit.entity;
+
+import gaji.service.domain.User;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class SearchKeyword {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private User user;
+
+    private String keyword;
+    private LocalDateTime searchDate;
+
+
+}
