+++ src/main/java/gaji/service/domain/recruit/entity/SelectCategory.java
@@ -1,6 +1,5 @@
-package gaji.service.domain.recruite;
+package gaji.service.domain.recruit.entity;
 
-import gaji.service.domain.User;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -9,18 +8,16 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class RecruitPostBookmark {
+public class SelectCategory {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
+    @ManyToOne(fetch =  FetchType.LAZY)
     @JoinColumn(name = "post_id")
     private RecruitPost recruitPost;
 
-
+    @ManyToOne(fetch =  FetchType.LAZY)
+    @JoinColumn(name = "category_id")
+    private Category category;
 }
