--- src/main/java/gaji/service/domain/recruite/SelectCategory.java
@@ -1,23 +1,0 @@
-package gaji.service.domain.recruite;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class SelectCategory {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch =  FetchType.LAZY)
-    @JoinColumn(name = "post_id")
-    private RecruitPost recruitPost;
-
-    @ManyToOne(fetch =  FetchType.LAZY)
-    @JoinColumn(name = "category_id")
-    private Category category;
-}
